refac: rename game-log/ -> game/
