+++ src/game/calPosition/score.position.ts
