+++ src/game/dto/game-record.dts.ts
