+++ src/game/dto/game-session.dto.ts
