+++ src/game/dto/game-socket.dto.ts
