+++ src/game/dto/ready.dto.ts
