+++ src/game/dto/rule.dto.ts
