+++ src/game/dto/spec.dto.ts
