+++ src/game/dto/status.dto.ts
