+++ src/game/game-log.service.ts
