+++ src/game/initializeGame/reset.ts
