+++ src/game/simulation.service.ts
